Rename truffle-environment to @truffle/environment
