+++ packages/truffle-core/lib/commands/console.js
@@ -21,7 +21,7 @@ const command = {
   run: function(options, done) {
     const Config = require("truffle-config");
     const Console = require("../console");
-    const { Environment } = require("truffle-environment");
+    const { Environment } = require("@truffle/environment");
 
     const config = Config.detect(options);
