+++ packages/truffle-core/lib/commands/debug.js
@@ -21,7 +21,7 @@ const command = {
     const debug = debugModule("lib:commands:debug");
 
     const Config = require("truffle-config");
-    const { Environment } = require("truffle-environment");
+    const { Environment } = require("@truffle/environment");
 
     const { CLIDebugger } = require("../debug");
