+++ packages/truffle-core/lib/commands/develop.js
@@ -24,7 +24,7 @@ const command = {
   },
   runConsole: (config, ganacheOptions, done) => {
     const Console = require("../console");
-    const { Environment } = require("truffle-environment");
+    const { Environment } = require("@truffle/environment");
 
     const commands = require("./index");
     const excluded = ["console", "develop", "unbox", "init"];
@@ -51,7 +51,7 @@ const command = {
   },
   run: (options, done) => {
     const Config = require("truffle-config");
-    const { Develop } = require("truffle-environment");
+    const { Develop } = require("@truffle/environment");
 
     const config = Config.detect(options);
     const customConfig = config.networks.develop || {};
