+++ packages/truffle/cli.webpack.config.js
@@ -19,7 +19,7 @@ module.exports = {
       __dirname,
       "../..",
       "node_modules",
-      "truffle-environment",
+      "@truffle/environment",
       "chain.js"
     ),
     analytics: path.join(
